requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP
spacy>=3.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM APIs
openai>=1.0.0
ollama>=0.1.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Model persistence
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
